Added product module (without validation)

+++ Cartify-Backend/src/main/java/in/mtk/cartify/repository/ProductRepo.java
@@ -4,6 +4,8 @@ import in.mtk.cartify.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.UUID;
+
 @Repository
-public interface ProductRepo extends JpaRepository<Product,Integer> {
+public interface ProductRepo extends JpaRepository<Product, Long> {
 }
